新增篩選條件

+++ ires-system/src/main/java/web/clinic/entity/Clinic.java
@@ -100,13 +100,13 @@ public class Clinic extends Core{
     private Timestamp updateTime;
 
     @Column(name = "latitude")
-    private BigDecimal latitude;
+    private Double latitude;
 
     @Column(name = "longitude")
-    private BigDecimal longitude;
+    private Double longitude;
     
     @Column(name = "rating")
-    private BigDecimal rating;
+    private Double rating;
 
     @Lob
     @Column(name = "profile_picture")
